+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         </activity>
         <activity android:name=".view.DetailActivity">
         </activity>
-        <activity android:name=".view.SearchActivity">
+        <activity android:name=".view.search.SearchActivity">
         </activity>
     </application>
