--- app/src/main/java/net/kwmt27/githubviewer/presenter/ISearchResultListPresenter.java
@@ -1,7 +1,0 @@
-package net.kwmt27.githubviewer.presenter;
-
-public interface ISearchResultListPresenter extends IBaseFragmentPresenter {
-
-
-
-}
