+++ app/src/main/java/net/kwmt27/githubviewer/presenter/search/ISearchResultListPresenter.java
@@ -1,0 +1,9 @@
+package net.kwmt27.githubviewer.presenter.search;
+
+import net.kwmt27.githubviewer.presenter.IBaseFragmentPresenter;
+
+public interface ISearchResultListPresenter extends IBaseFragmentPresenter {
+
+
+
+}
